Completion of Normalization
Descriptions and normalization complete. In truth, I'm unsure on whether
the 2NF is correct. I feel I know the outcome of the database, and that
it will be the same result as I have proposed here. However, I'm not
sure if my process is entirely correct. I have justified my choices, but
again, I'm not sure how accurate they are towards that actual
normalization validation rules, even while I understand the meaning of
the rules.