Added SQL Database Typescript
Completed structure of SQL Database, creating tables, drop tables and
inserting sample data. Completed most queries. A little stuck on the
last one.
I can get the data in UID form, I can get the names of the Lifts and
Pistes (exactly what I need), but only separately. I need a way to join
the two of them, just like how they are when displayed in this UID
query.
Ah, and I should add, I'm now using a UID for the Primary Key, keeping
the names of the Pistes and Lifts UNIQUE in order to constrain the
Candidate Key. Part of me is very tempted to just straight up make the
Names the Primary Key. This would imply that we never expect a user to
update or change the name though...
Either way, my documentation needs another overhaul to reflect this.
I'll sleep on it and decide whether to keep using the UID or to depend
on the name alone in the morning.