Put lift/piste_name into the M-M for content
Started to update the documentation.
Updated the SQL to correctly create, insert and select data from the
database now that piste and lift name are both in connections.
I chose this way for the use of content. I will explain it in my
documentation.
It could be removed, and just list the UIDs, and then translate those
into the names elsewhere, but idk how to do it right now, and it seems
to make more content sense for a user when looking at the connections
table to see the names. I'll do my best to justify this within my
documentation.